requirements.txt
Pillow==10.0.1
easyocr==1.7.1
streamlit==1.27.2
opencv-python==4.9.0.80
numpy==1.24.3
imageio==2.13.1
